+++ src/plugins/index.js
@@ -14,8 +14,8 @@ export function usePlugins(app) {
   });
   app.use(polkadot, {
     start: false,
-    // endpoint: "wss://kusama.rpc.robonomics.network/",
-    endpoint: "wss://polkadot.rpc.robonomics.network/",
+    endpoint: "wss://kusama.rpc.robonomics.network/",
+    // endpoint: "wss://polkadot.rpc.robonomics.network/",
   });
   useI18n(app);
   useIcons(app);